+++ kubejs/client_scripts/Tooltips.js
@@ -8,4 +8,6 @@ ItemEvents.tooltip(event => {
 			}
 		}
 	)
+
+	event.add(['minecraft:furnace', 'minecraft:blast_furnace', 'minecraft:smoker'], Text.yellow('No functionality.').append(Text.gray(' Used for crafting or decoration only.')))
 })